requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
dotenv-python==0.0.1
Jinja2==3.1.2
matplotlib==3.5.1
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-repackaged==0.3.1
mysqlclient==2.1.1
pandas==1.4.0
pdfkit==1.0.0
plotly==5.10.0
PyPDF2==2.10.7
reportlab==3.6.11
requests==2.25.1

